Move variations module into fontdrasil
