+++ fontdrasil/src/lib.rs
@@ -7,3 +7,4 @@ pub mod paths;
 mod piecewise_linear_map;
 pub mod types;
 pub mod util;
+pub mod variations;
